// frontend/src/app/core/generated/circabc/model/news.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Node } from './node';


export interface News { 
    id?: string;
    /**
     * The object that is used to compile all the translations of a node into a JSON object. It is basically composed of a map with a key languaguage code and its value 
     */
    title?: { [key: string]: string; };
    content?: string;
    date?: string;
    pattern?: News.PatternEnum;
    layout?: News.LayoutEnum;
    /**
     * valid values are  1 / 2 / 3 
     */
    size?: number;
    files?: Array<Node>;
    url?: string;
    modified?: string;
    modifier?: string;
    created?: string;
    creator?: string;
    /**
     * Representation of the permissions given to the user making the call. To know what are the permissions for each services inside one group. 
     */
    permissions?: { [key: string]: string; };
    properties?: { [key: string]: string; };
}
export namespace News {
    export const PatternEnum = {
        Text: 'text',
        Document: 'document',
        Image: 'image',
        Date: 'date',
        Iframe: 'iframe'
    } as const;
    export type PatternEnum = typeof PatternEnum[keyof typeof PatternEnum];
    export const LayoutEnum = {
        Normal: 'normal',
        Important: 'important',
        Reminder: 'reminder'
    } as const;
    export type LayoutEnum = typeof LayoutEnum[keyof typeof LayoutEnum];
}